+++ src/manager.ts
@@ -1,12 +1,15 @@
+import path = require("path");
 import * as vscode from "vscode";
+import { getImageAssetsFromFolder } from "./assets";
 import { Logger, LogLevel } from "./logging";
+import { readManifestFile } from "./manifest";
 import { getErrorMessage } from "./utils";
 
 export class ExtensionManager implements vscode.Disposable {
   private _context: vscode.ExtensionContext;
   private _channel: vscode.OutputChannel;
   public logger: Logger = new Logger();
-  private _terminal: vscode.Terminal | null = null;
+  private _isRaycastEnabled = false;
 
   constructor(public readonly extensionContext: vscode.ExtensionContext) {
     this._context = extensionContext;
@@ -20,11 +23,80 @@ export class ExtensionManager implements vscode.Disposable {
         await this.updateState();
       })
     );
+    this.registerPackageJsonChanges();
+    this.registerCompletionProviders();
+  }
+
+  private registerPackageJsonChanges() {
+    const triggerUpdateOnChange = (filenames: string[]) => {
+      const pkgjson = this.getActiveWorkspacePackageFilename();
+      if (pkgjson) {
+        const found = filenames.find((f) => f === pkgjson);
+        if (found) {
+          this.updateState();
+        }
+      }
+    };
+    this._context.subscriptions.push(
+      vscode.workspace.onDidCreateFiles((e) => {
+        triggerUpdateOnChange(e.files.map((f) => f.fsPath));
+      })
+    );
+    this._context.subscriptions.push(
+      vscode.workspace.onDidRenameFiles((e) => {
+        const files: string[] = [];
+        e.files.forEach((f) => {
+          files.push(f.newUri.fsPath);
+          files.push(f.oldUri.fsPath);
+        });
+        triggerUpdateOnChange(files);
+      })
+    );
+    this._context.subscriptions.push(
+      vscode.workspace.onDidDeleteFiles((e) => {
+        triggerUpdateOnChange(e.files.map((f) => f.fsPath));
+      })
+    );
+    this._context.subscriptions.push(
+      vscode.workspace.onDidSaveTextDocument((e) => {
+        const pkgjson = this.getActiveWorkspacePackageFilename();
+        if (pkgjson && pkgjson === e.fileName) {
+          this.updateState();
+        }
+      })
+    );
   }
 
   public async updateState(): Promise<void> {
     this.logger.level = this.getLogLevel();
     this.logger.debug("update state");
+    await this.updateContext();
+  }
+
+  public async updateContext(): Promise<void> {
+    const pkgjsonFilename = this.getActiveWorkspacePackageFilename();
+    const manifest = await readManifestFile(pkgjsonFilename);
+    let isRaycastEnabled = false;
+    if (manifest) {
+      const deps = manifest.dependencies;
+      if (deps) {
+        const raycastapi: string | undefined = deps["@raycast/api"];
+        if (raycastapi) {
+          isRaycastEnabled = true;
+        }
+      }
+    }
+    this._isRaycastEnabled = isRaycastEnabled;
+    this.logger.debug(`${pkgjsonFilename} raycast enabled: ${isRaycastEnabled}`);
+    await this.setContext("raycast.workspaceEnabled", isRaycastEnabled);
+  }
+
+  private async setContext(key: string, value: any): Promise<void> {
+    try {
+      this.logger.debug(`setContext ${key} to ${value}`);
+      await vscode.commands.executeCommand("setContext", key, value);
+      this.logger.debug("setContent succeeded");
+    } catch (error) {}
   }
 
   public registerCommand(command: string, callback: (...args: any[]) => any, thisArg?: any): vscode.Disposable {
@@ -89,6 +161,10 @@ export class ExtensionManager implements vscode.Disposable {
     return result;
   }
 
+  get isRaycastEnabled(): boolean {
+    return this._isRaycastEnabled;
+  }
+
   public runNpmExec(cmd: string[], terminalID?: string | undefined) {
     const term = this.getTerminal(terminalID);
     term.sendText("clear");
@@ -110,6 +186,91 @@ export class ExtensionManager implements vscode.Disposable {
       return ws;
     }
     return undefined;
+  }
+
+  public getActiveWorkspacePackageFilename(): string | undefined {
+    const ws = this.getActiveWorkspace();
+    if (ws) {
+      return path.join(ws.uri.fsPath, "package.json");
+    }
+    return undefined;
+  }
+
+  public async getImageAssets(): Promise<string[]> {
+    const ws = this.getActiveWorkspace();
+    if (!ws || !this.isRaycastEnabled) {
+      return [];
+    }
+    const assetsFolder = path.join(ws.uri.fsPath, "assets");
+    return await getImageAssetsFromFolder(assetsFolder);
+  }
+
+  private registerCompletionProviders() {
+    const self = this;
+    const tsImageAssetCompletionProvider = vscode.languages.registerCompletionItemProvider(
+      "typescriptreact",
+      {
+        async provideCompletionItems(
+          document: vscode.TextDocument,
+          position: vscode.Position,
+          token: vscode.CancellationToken,
+          context: vscode.CompletionContext
+        ) {
+          const line = document.lineAt(position.line);
+          const text = line.text.substring(0, position.character);
+          const sourceIndex = text.lastIndexOf("source:");
+          const iconIndex = text.lastIndexOf("icon=");
+          if ((sourceIndex > 0 && text.length - sourceIndex < 15) || (iconIndex > 0 && text.length - iconIndex < 15)) {
+            const assets = await self.getImageAssets();
+            if (assets && assets.length > 0) {
+              return assets.map((a) => new vscode.CompletionItem(a, vscode.CompletionItemKind.File));
+            }
+          }
+          return undefined;
+        },
+      },
+      '"',
+      "'"
+    );
+    const jsonImageAssetCompletionProvider = vscode.languages.registerCompletionItemProvider(
+      "json",
+      {
+        async provideCompletionItems(
+          document: vscode.TextDocument,
+          position: vscode.Position,
+          token: vscode.CancellationToken,
+          context: vscode.CompletionContext
+        ) {
+          const filename = path.basename(document.fileName);
+          if (filename !== "package.json") {
+            return undefined;
+          }
+          const line = document.lineAt(position.line);
+          const text = line.text.substring(0, position.character);
+          const lastColon = text.lastIndexOf(":");
+          if (lastColon > 0) {
+            const splits = text.substring(0, lastColon).split(":");
+            if (splits && splits.length > 0) {
+              const last = splits[splits.length - 1];
+              if (last.includes('"icon"')) {
+                const assets = await self.getImageAssets();
+                if (assets && assets.length > 0) {
+                  const comps = assets.map((a) => {
+                    const c = new vscode.CompletionItem(a, vscode.CompletionItemKind.File);
+                    c.range = new vscode.Range(position, position); // required for json files
+                    return c;
+                  });
+                  return comps;
+                }
+              }
+            }
+          }
+          return undefined;
+        },
+      },
+      '"'
+    );
+    this._context.subscriptions.push(tsImageAssetCompletionProvider, jsonImageAssetCompletionProvider);
   }
 
   public dispose() {}