+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.3
+
+- Add `Insert Image Asset`
+- Add Code snippets prefix with `raycast`
+- Add Image Asset autocompletion for extension icons in `.tsx` files (`source:` and `icon=` are supported)
+- Add Image Asset autocompletion for extension icons in `package.json` (`"icon":` is supported)
+
 ## 0.2
 
 - Add command to add a Raycast Preference